requirements.txt
streamlit
torch
torchvision
opencv-python
pillow
numpy
matplotlib
scikit-learn
tqdm
